scripts/python/requirements.txt
mysql-connector~=2.2.9
requests>=2.18.0
configparser~=4.0.2